requirements.txt
# Core
aiohttp>=3.8.0
python-dotenv>=1.0.0
requests>=2.28.0

# MongoDB
pymongo>=4.6.0
dnspython>=2.4.0

# YouTube Scraping
scrapetube>=2.5.0
yt-dlp>=2023.12.0

# AI / Bedrock
boto3>=1.34.0

# CLI & Display
tabulate>=0.9.0

# Email
secure-smtplib>=0.1.1
